% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{branching_sample}
\alias{branching_sample}
\title{Resample particle indices}
\usage{
branching_sample(n_part, norm_weights)
}
\arguments{
\item{n_part}{Total number of particles, a positive integer.}

\item{norm_weights}{Normalized weights for all particles, a vector of length
\code{n_part}.}
}
\value{
A vector of length \code{n_part}; the element in \code{i}-th position
corresponds to the number of times the particle index \code{i} was resampled.
}
\description{
This function resamples particle \emph{indices} using sampling with replacement.
}
