% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{stochSIR_likelihood_weights}
\alias{stochSIR_likelihood_weights}
\title{Compute likelihood weights}
\usage{
stochSIR_likelihood_weights(states, current_obs, fixed_inputs)
}
\arguments{
\item{states}{A data frame with columns \code{psi}, \code{I} and \code{R}
(corresponding to \eqn{\psi, i} and \eqn{r}, respectively).}

\item{current_obs}{Number of people detected at the current time point, a
positive integer.}

\item{fixed_inputs}{Necessary arguments to further specify the chosen
transition kernel, a list. Given the updating schemes describes in the paper,
it should include the following:
\itemize{
\item \code{tot_pop}: total number of individuals in the considered population
(indicated by \eqn{N} in the paper), a positive integer;
\item \code{g}: inverse of the average time a non-detected individual stays
infectious (indicated by \eqn{\gamma} in the paper), a positive scalar;
\item \code{q}: probability of detecting an infectious person through testing
(indicated by \eqn{q} in the paper), a non-negative scalar;
\item \code{d_f}: inverse of the average time an infected person enjoys immunity
infectious (indicated by \eqn{\delta} in the paper), a positive scalar;
\item \code{quarantine}: logical; if TRUE, the transition kernel corresponding
to the model with quarantine is used.
}
Defaults to \code{stochSIR_state_fixed_inputs}, which is available to users upon
package loading.}
}
\value{
A list containing \code{state_norm_weights} (normalized state weights)
and \code{param_weight} (unnormalized parameter weight for the parameter vector
corresponding to the \code{states} data frame passed as argument).
}
\description{
This function computes the \emph{normalized} state weights and the \emph{unnormalized}
parameter weight for a given parameter vector and its corresponding states particles.
}
