% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPF.R
\name{stochSIR_NPF}
\alias{stochSIR_NPF}
\title{Obtain posterior estimates through NPF algorithm}
\usage{
stochSIR_NPF(
  n_params = 500,
  n_state = 500,
  param_prior_settings = stochSIR_param_prior_settings,
  state_prior_settings = stochSIR_state_prior_settings,
  jitter_settings = stochSIR_jitter_settings,
  fixed_inputs = stochSIR_fixed_inputs,
  tot_time_steps = length(observations),
  delta = 1,
  observations = covid_AT$rollmean
)
}
\arguments{
\item{n_params}{Number of parameter particles to sample, a positive scalar. Defaults to 500.}

\item{n_state}{Number of state particles to sample, a positive scalar. Defaults to 500.}

\item{param_prior_settings}{Necessary arguments to further specify the chosen
prior distribution for each parameter, a (nested) list. The top-level names must
be \code{k_settings}, \code{s_settings} and \code{m_settings}. In the paper,
the prior distribution of each parameter is chosen to be uniform, so each
top-level element must be a list containing \code{min} and \code{max}. Defaults
to \code{stochSIR_param_prior_settings}, which is available to users upon
package loading.}

\item{state_prior_settings}{Necessary arguments to further specify the chosen
prior distribution for each state variable, a (nested) list. The top-level
names must be \code{psi_settings} and \code{I_settings}. In the paper, the prior
distribution of \eqn{\psi} is chosen to be normal and the one of \eqn{I} to be
gamma, so the corresponding top-level elements must be a list containing \code{mean}
and \code{sd} and a list containing \code{rate} and \code{shape}. Note that the
paper assumes no-one has recovered at start time, so \eqn{R} is initialized at
0 for all particles and requires no further specification settings. Defaults to
stochSIR_state_prior_settings, which is available to users upon package loading.}

\item{jitter_settings}{Necessary arguments to further specify the chosen
jittering kernel, a (nested) list. The top-level names must be \code{k_settings},
\code{s_settings} and \code{m_settings}. In the paper, the jittering kernel is
chosen to be a truncated Gaussian one and this function calls
\code{truncnorm::rtruncnorm}, so each top-level element must be a list containing
\code{a}, \code{b} and \code{max}. Defaults to \code{stochSIR_jitter_settings},
which is available to users upon package loading.}

\item{fixed_inputs}{Necessary arguments to further specify the chosen
transition kernel, a list. Given the updating schemes describes in the paper,
it should include the following:
\itemize{
\item \code{tot_pop}: total number of individuals in the considered population
(indicated by \eqn{N} in the paper), a positive integer;
\item \code{g}: inverse of the average time a non-detected individual stays
infectious (indicated by \eqn{\gamma} in the paper), a positive scalar;
\item \code{q}: probability of detecting an infectious person through testing
(indicated by \eqn{q} in the paper), a non-negative scalar;
\item \code{d_f}: inverse of the average time an infected person enjoys immunity
infectious (indicated by \eqn{\delta} in the paper), a positive scalar;
\item \code{quarantine}: logical; if TRUE, the transition kernel corresponding
to the model with quarantine is used.
}
Defaults to \code{stochSIR_state_fixed_inputs}, which is available to users upon
package loading.}

\item{tot_time_steps}{Total number of time steps at which to obtain posterior
estimates, a positive integer. Defaults to \code{length{observations}}.}

\item{delta}{Time-step size, a positive scalar. Defaults to 1.}

\item{observations}{(Univariate) time series of observations, a numeric vector.
Defaults to \code{covid_AT$rollmean} (data set \code{covid_AT} is available
to users upon package loading).}
}
\description{
This function runs the Nested Particle Filter (NPF) algorithm of Crisan and
Míguez (2018) to obtain posterior estimates for states and parameters at each
time point. The function defaults are such that the model of Colaneri et al.
(2025) is assumed.
}
