% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{stochSIR_NPF_output}
\alias{stochSIR_NPF_output}
\title{Compute and format relevant NPF outputs}
\usage{
stochSIR_NPF_output(state_particles, param_particles, fixed_inputs)
}
\arguments{
\item{state_particles}{List of state particles, each element of which is a data
frame with columns \code{psi}, \code{I} and \code{R} (corresponding to \eqn{\psi, i}
and \eqn{r}, respectively).}

\item{param_particles}{Parameter particles, a data frame
with columns \code{k}, \code{s} and \code{m}.}

\item{fixed_inputs}{Necessary arguments to further specify the chosen
transition kernel, a list. Given the updating schemes describes in the paper,
it should include the following:
\itemize{
\item \code{tot_pop}: total number of individuals in the considered population
(indicated by \eqn{N} in the paper), a positive integer;
\item \code{g}: inverse of the average time a non-detected individual stays
infectious (indicated by \eqn{\gamma} in the paper), a positive scalar;
\item \code{q}: probability of detecting an infectious person through testing
(indicated by \eqn{q} in the paper), a non-negative scalar;
\item \code{d_f}: inverse of the average time an infected person enjoys immunity
infectious (indicated by \eqn{\delta} in the paper), a positive scalar;
\item \code{quarantine}: logical; if TRUE, the transition kernel corresponding
to the model with quarantine is used.
}
Defaults to \code{stochSIR_state_fixed_inputs}, which is available to users upon
package loading.}
}
\value{
A list with top-level names \code{rep_rate_res}, \code{k_res},
\code{s_res}, \code{m_res} and \code{ratio_res} (corresponding to the reproduction
number \eqn{\mathcal{R}}, the parameters \eqn{\kappa}, \eqn{\sigma}, \eqn{\mu}
and the ratio \eqn{\sigma^2/2\kappa}, respectively). Each top-level element is
a data frame containing the mean, the 5\% and the 95\% quantile of the posterior
distribution of each variable of interest.
}
\description{
This function computes relevant characteristics of the approximate posterior
distribution and formats them 'by variable'.
}
