% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{stochSIR_param_prior}
\alias{stochSIR_param_prior}
\title{Sample parameter particles from their prior distribution}
\usage{
stochSIR_param_prior(
  n_params,
  param_prior_settings = stochSIR_param_prior_settings
)
}
\arguments{
\item{n_params}{Number of parameter particles to sample, a positive scalar.}

\item{param_prior_settings}{Necessary arguments to further specify the chosen
prior distribution for each parameter, a (nested) list. The top-level names must
be \code{k_settings}, \code{s_settings} and \code{m_settings}. In the paper,
the prior distribution of each parameter is chosen to be uniform, so each
top-level element must be a list containing \code{min} and \code{max}. Defaults
to \code{stochSIR_param_prior_settings}, which is available to users upon
package loading.}
}
\value{
A data frame with \code{n_params} rows and columns \code{k}, \code{s}
and \code{m} (corresponding to \eqn{\kappa, \sigma} and \eqn{\mu}, respectively).
}
\description{
This function samples parameter particles \eqn{\theta = (\kappa, \sigma, \mu)}
from their prior distribution.
}
