% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{stochSIR_state_prior}
\alias{stochSIR_state_prior}
\title{Sample state particles from their prior distribution}
\usage{
stochSIR_state_prior(
  n_state,
  state_prior_settings = stochSIR_state_prior_settings
)
}
\arguments{
\item{n_state}{Number of state particles to sample, a positive scalar.}

\item{state_prior_settings}{Necessary arguments to further specify the chosen
prior distribution for each state variable, a (nested) list. The top-level
names must be \code{psi_settings} and \code{I_settings}. In the paper, the prior
distribution of \eqn{\psi} is chosen to be normal and the one of \eqn{I} to be
gamma, so the corresponding top-level elements must be a list containing \code{mean}
and \code{sd} and a list containing \code{rate} and \code{shape}. Note that the
paper assumes no-one has recovered at start time, so \eqn{R} is initialized at
0 for all particles and requires no further specification settings. Defaults to
stochSIR_state_prior_settings, which is available to users upon package loading.}
}
\value{
A data frame with \code{n_state} rows and columns \code{psi}, \code{I}
and \code{R} (corresponding to \eqn{\psi, i} and \eqn{r}, respectively).
}
\description{
This function samples state particles \eqn{x = (\psi, i, r)}
from their prior distribution.
}
