% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{stochSIR_state_step}
\alias{stochSIR_state_step}
\title{Propagate state particles from their transition kernel}
\usage{
stochSIR_state_step(
  states,
  params,
  fixed_inputs = stochSIR_fixed_inputs,
  delta,
  previous_obs
)
}
\arguments{
\item{states}{A data frame with columns \code{psi}, \code{I} and \code{R}
(corresponding to \eqn{\psi, i} and \eqn{r}, respectively).}

\item{params}{A named vector with elements \code{k}, \code{m} and \code{s}
(corresponding to \eqn{\kappa, \sigma} and \eqn{\mu}, respectively).}

\item{fixed_inputs}{Necessary arguments to further specify the chosen
transition kernel, a list. Given the updating schemes describes in the paper,
it should include the following:
\itemize{
\item \code{tot_pop}: total number of individuals in the considered population
(indicated by \eqn{N} in the paper), a positive integer;
\item \code{g}: inverse of the average time a non-detected individual stays
infectious (indicated by \eqn{\gamma} in the paper), a positive scalar;
\item \code{q}: probability of detecting an infectious person through testing
(indicated by \eqn{q} in the paper), a non-negative scalar;
\item \code{d_f}: inverse of the average time an infected person enjoys immunity
infectious (indicated by \eqn{\delta} in the paper), a positive scalar;
\item \code{quarantine}: logical; if TRUE, the transition kernel corresponding
to the model with quarantine is used.
}
Defaults to \code{stochSIR_state_fixed_inputs}, which is available to users upon
package loading.}

\item{delta}{Time-step size, a positive scalar.}

\item{previous_obs}{Number of people detected at the previous time point and
removed from pool of infectious people, a positive integer. Only used if
\code{quarantine} is \code{TRUE}.}
}
\value{
A data frame with the same number of rows as argument \code{states} and
columns \code{psi}, \code{I} and \code{R} (corresponding to \eqn{\psi, i} and
\eqn{r}, respectively).
}
\description{
For a given parameter particle, this function propagates state particles \eqn{x = (\psi, i, r)}
from their transition kernel.
}
